// FacetsUI/src/app/modules/team-member/models/create-team-member.model.ts
import { Title } from "src/app/core/enums/title.enum";
import { VisitorIdentityType } from "src/app/core/enums/visitor-identityType.enum";

export class CreateTeamMemberModel {
    title: Title;
    identityType: VisitorIdentityType
    nICNumber: string;
    passportNumber: string;
    countryId: string;
    firstName: string;
    lastName: string;
    eventId: string;
    mobileNumber: string;
    email: string;
    imageURL: string = '../../assets/images/no-image.png';
    address = new Address();
    companyName: string | null;
}

export class Address {
    address: string;
}